Update from mirror (#1112)
* Faster tests

* Add github workflow

* Faster test

* Fix MPI dependency

* Faster HER tests + fix CI

* No specific python version for pytype

* Fixes + add badge

* Fix multiprocessing error

* Separate TD3 test

* Add tolerance for deterministic check

* Better tests for saving/loading

* Remove unnecessary check

* Add comment about VecEnv start method

* Make pytype happy

* Debug MPI

* Add MPI step

* Move MPI tests outside pytest

* Two processes for GitHub CI

* Deactivate check moments

* Fix import error

* Copy version.txt to docker container (#2)

* Copy version.txt to docker container

* Update changelog

* Update Max username


* Fixed typo (#3)

* Fixed typo

* Update changelog.rst


* Warn users to switch to Stable Baselines3 (#4)

* Warn users to switch to Stable Baselines3

* Fix for pytype

* Fix other pytype error

* Clarify update warning

* Remove python 3.5 build

* Allow test to fail

* Flaky TD3 test

* Fix argument

+++ tests/test_continuous.py
@@ -16,7 +16,7 @@ from tests.test_common import _assert_eq
 
 
 N_EVAL_EPISODES = 20
-NUM_TIMESTEPS = 15000
+NUM_TIMESTEPS = 300
 
 MODEL_LIST = [
     A2C,
@@ -40,6 +40,7 @@ def test_model_manipulation(request, model_class):
 
     :param model_class: (BaseRLModel) A model
     """
+    model_fname = None
     try:
         env = DummyVecEnv([lambda: IdentityEnvBox(eps=0.5)])
 
@@ -47,7 +48,9 @@ def test_model_manipulation(request, model_class):
         model = model_class(policy="MlpPolicy", env=env, seed=0)
         model.learn(total_timesteps=NUM_TIMESTEPS)
 
-        acc_reward, _ = evaluate_policy(model, env, n_eval_episodes=N_EVAL_EPISODES)
+        env.reset()
+        observations = np.concatenate([env.step([env.action_space.sample()])[0] for _ in range(10)], axis=0)
+        selected_actions, _ = model.predict(observations, deterministic=True)
 
         # saving
         model_fname = './test_model_{}.zip'.format(request.node.name)
@@ -58,11 +61,13 @@ def test_model_manipulation(request, model_class):
         # loading
         model = model_class.load(model_fname)
 
+        # check if model still selects the same actions
+        new_selected_actions, _ = model.predict(observations, deterministic=True)
+        assert np.allclose(selected_actions, new_selected_actions, 1e-4)
+
         # changing environment (note: this can be done at loading)
         env = DummyVecEnv([lambda: IdentityEnvBox(eps=0.5)])
         model.set_env(env)
-
-        loaded_acc_reward, _ = evaluate_policy(model, env, n_eval_episodes=N_EVAL_EPISODES)
 
         obs = env.reset()
         with pytest.warns(None) as record:
@@ -95,43 +100,22 @@ def test_model_manipulation(request, model_class):
             actions_logprobas = model.action_probability(observations, actions=actions, logp=True)
             assert np.allclose(actions_probas, np.exp(actions_logprobas)), (actions_probas, actions_logprobas)
 
-        # assert <15% diff
-        assert abs(acc_reward - loaded_acc_reward) / max(acc_reward, loaded_acc_reward) < 0.15, \
-            "Error: the prediction seems to have changed between loading and saving"
-
         # learn post loading
         model.learn(total_timesteps=100)
 
-        # validate no reset post learning
-        # This test was failing from time to time for no good reason
-        # other than bad luck
-        # We should change this test
-        # loaded_acc_reward = 0
-        # set_global_seeds(0)
-        # obs = env.reset()
-        # for _ in range(N_EVAL_EPISODES):
-        #     action, _ = model.predict(obs)
-        #     obs, reward, _, _ = env.step(action)
-        #     loaded_acc_reward += reward
-        # loaded_acc_reward = sum(loaded_acc_reward) / N_EVAL_EPISODES
-        # # assert <10% diff
-        # assert abs(acc_reward - loaded_acc_reward) / max(acc_reward, loaded_acc_reward) < 0.1, \
-        #     "Error: the prediction seems to have changed between pre learning and post learning"
-
         # predict new values
-
         evaluate_policy(model, env, n_eval_episodes=N_EVAL_EPISODES)
 
         # Free memory
         del model, env
 
     finally:
-        if os.path.exists(model_fname):
+        if model_fname is not None and os.path.exists(model_fname):
             os.remove(model_fname)
 
 
 def test_ddpg():
-    args = ['--env-id', 'Pendulum-v0', '--num-timesteps', 1000, '--noise-type', 'ou_0.01']
+    args = ['--env-id', 'Pendulum-v0', '--num-timesteps', 300, '--noise-type', 'ou_0.01']
     args = list(map(str, args))
     return_code = subprocess.call(['python', '-m', 'stable_baselines.ddpg.main'] + args)
     _assert_eq(return_code, 0)
@@ -146,7 +130,7 @@ def test_ddpg_eval_env():
     model = DDPG("MlpPolicy", "Pendulum-v0", nb_rollout_steps=5,
                 nb_train_steps=2, nb_eval_steps=10,
                 eval_env=eval_env, verbose=0)
-    model.learn(1000)
+    model.learn(NUM_TIMESTEPS)
 
 
 def test_ddpg_normalization():
@@ -157,7 +141,7 @@ def test_ddpg_normalization():
     model = DDPG('MlpPolicy', 'Pendulum-v0', memory_limit=50000, normalize_observations=True,
                  normalize_returns=True, nb_rollout_steps=128, nb_train_steps=1,
                  batch_size=64, param_noise=param_noise)
-    model.learn(1000)
+    model.learn(NUM_TIMESTEPS)
     obs_rms_params = model.sess.run(model.obs_rms_params)
     ret_rms_params = model.sess.run(model.ret_rms_params)
     model.save('./test_ddpg.zip')
@@ -185,4 +169,4 @@ def test_ddpg_popart():
     model = DDPG('MlpPolicy', 'Pendulum-v0', memory_limit=50000, normalize_observations=True,
                  normalize_returns=True, nb_rollout_steps=128, nb_train_steps=1,
                  batch_size=64, action_noise=action_noise, enable_popart=True)
-    model.learn(1000)
+    model.learn(NUM_TIMESTEPS)
